# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Daemon Host Panel
  version: 0.1.0
paths:
...
  /api/participant/wordcloud/word:
    post:
      tags:
      - wordcloud
      summary: Submit Word
      description: Participant submits a word to the word cloud.
      operationId: submit_word_api_participant_wordcloud_word_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitWordBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
      x-feature: wordcloud
  /api/{session_id}/host/wordcloud/word:
    post:
      tags:
      - wordcloud
      summary: Host Submit Word
      description: Host submits a word — same as participant but no scoring.
      operationId: host_submit_word_api__session_id__host_wordcloud_word_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitWordBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
      x-feature: wordcloud
  /api/{session_id}/host/wordcloud/topic:
    post:
      tags:
      - wordcloud
      summary: Set Topic
      description: Host sets the word cloud topic.
      operationId: set_topic_api__session_id__host_wordcloud_topic_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetTopicBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
      x-feature: wordcloud
...
  /api/{session_id}/host/quiz-request:
    post:
      tags:
      - quiz
      summary: Request Quiz
      description: Host requests a quiz — stores request for the orchestrator loop
        to pick up.
      operationId: request_quiz_api__session_id__host_quiz_request_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizRequestBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
      x-feature: quiz
...
  /api/{session_id}/host/quiz-refine:
    post:
      tags:
      - quiz
      summary: Request Quiz Refine
      description: Host requests regeneration of a specific question or option.
      operationId: request_quiz_refine_api__session_id__host_quiz_refine_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuizRefineRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
      x-feature: quiz
...
  /api/session/start:
    post:
      tags:
      - session
      summary: Start Session
      description: Host starts a new session (creates folder, assigns session_id,
        clean slate).
      operationId: start_session_api_session_start_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartSessionRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStartResponse'
      x-feature: session_management
...
  /api/session/resume:
    post:
      tags:
      - session
      summary: Resume Session
      description: Host resumes an existing session folder. Uses session-state.json
        as persisted storage.
      operationId: resume_session_api_session_resume_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResumeSessionRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionStartResponse'
      x-feature: session_management
...
  /api/{session_id}/session/interval-lines.txt:
    get:
      tags:
      - session
      summary: Get Interval Lines Txt
      description: Return raw transcript lines for a time window.
      operationId: get_interval_lines_txt_api__session_id__session_interval_lines_txt_get
      parameters:
      - name: start
        in: query
        required: true
        schema:
          type: string
          description: Interval start in ISO format
          title: Start
        description: Interval start in ISO format
      - name: end
        in: query
        required: true
        schema:
          type: string
          description: Interval end in ISO format
          title: End
        description: Interval end in ISO format
      responses:
        '200':
          description: Successful Response
          content:
            text/plain:
              schema:
                type: string
      x-feature: session_management
      x-doc-notes:
      - Returns text/plain interval lines for session export/inspection.
...
components:
  schemas:
...
    OkResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
      type: object
      title: OkResponse
...
    QuizPreviewPayload-Input:
      properties:
        quiz:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Quiz
        question:
          anyOf:
          - type: string
          - type: 'null'
          title: Question
        options:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Options
        multi:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Multi
        correct_indices:
          anyOf:
          - items:
              type: integer
            type: array
          - type: 'null'
          title: Correct Indices
      type: object
      title: QuizPreviewPayload
...
    QuizRefineRequest:
      properties:
        target:
          type: string
          title: Target
        preview:
          anyOf:
          - $ref: '#/components/schemas/QuizPreviewPayload-Input'
          - type: 'null'
      type: object
      required:
      - target
      title: QuizRefineRequest
    QuizRequestBody:
      properties:
        minutes:
          anyOf:
          - type: integer
          - type: 'null'
          title: Minutes
        topic:
          anyOf:
          - type: string
          - type: 'null'
          title: Topic
      type: object
      title: QuizRequestBody
...
    ResumeSessionRequest:
      properties:
        folder:
          type: string
          title: Folder
      type: object
      required:
      - folder
      title: ResumeSessionRequest
...
    SessionStartResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
        session_name:
          type: string
          title: Session Name
        session_id:
          type: string
          title: Session Id
      type: object
      required:
      - session_name
      - session_id
      title: SessionStartResponse
...
    SetTopicBody:
      properties:
        topic:
          type: string
          title: Topic
      type: object
      required:
      - topic
      title: SetTopicBody
...
    StartSessionRequest:
      properties:
        name:
          type: string
          title: Name
        type:
          type: string
          title: Type
          default: workshop
      type: object
      required:
      - name
      title: StartSessionRequest
...
    SubmitWordBody:
      properties:
        word:
          type: string
          title: Word
      type: object
      required:
      - word
      title: SubmitWordBody
...